More work towards allowing the container sync to explicitly called by the client.
