+++ Microsoft.Azure.CosmosRepository/src/Services/ICosmosContainerSyncService.cs
@@ -1,0 +1,21 @@
+﻿// Copyright (c) IEvangelist. All rights reserved.
+// Licensed under the MIT License.
+
+using System.Threading.Tasks;
+
+namespace Microsoft.Azure.CosmosRepository.Services
+{
+    /// <summary>
+    /// Allows containers properties to be sunk.
+    /// </summary>
+    public interface ICosmosContainerSyncService
+    {
+
+        /// <summary>
+        /// Syncs a specific container for an <see cref="IItem"/>
+        /// </summary>
+        /// <typeparam name="TItem">The <see cref="IItem"/> to sync.</typeparam>
+        /// <returns></returns>
+        Task SyncAsync<TItem>() where TItem : IItem;
+    }
+}
